requirements.txt
streamlit
pandas
plotly
google-generativeai
pillow
supabase